Fix manual merge conflicts by merge 9ce50db into 91c3078
