+++ evaluator/evaluator/app/core/dataset_handler.py
@@ -4,9 +4,8 @@ from typing import Union
 
 import datasets
 from datasets import Dataset, DatasetDict, DownloadMode, Split
+from evaluator.app.core.dataset_formatter import DatasetFormatter
 from evaluator.app.settings import DatasetHandlerSettings
-from evaluator.core.dataset_formatter import DatasetFormatter
-from evaluator.settings import DatasetHandlerSettings
 
 logger = logging.getLogger(__name__)
