--- evaluator/evaluator/routers/api.py
@@ -1,11 +1,0 @@
-from fastapi import APIRouter
-
-from evaluator.routers.evaluator import router as evaluator_router
-from evaluator.routers.example import router as example_router
-from evaluator.routers.predictor import router as predictor_router
-
-router = APIRouter()
-
-router.include_router(example_router)
-router.include_router(evaluator_router)
-router.include_router(predictor_router)
